refactor(sync-actions): move category custom actions to utils
